Bug 1232306: cast char index to uint8_t explicitly to suppress compiler warning "array subscript is of type 'char' [-Wchar-subscripts]". r=rkent

+++ mailnews/import/src/ImportCharSet.cpp
@@ -40,19 +40,19 @@ UInitMaps::UInitMaps()
   ImportCharSet::m_Ascii[ImportCharSet::cLinefeedChar] |= ImportCharSet::cWhiteSpaceChar;
   ImportCharSet::m_Ascii[ImportCharSet::cSpaceChar] |= ImportCharSet::cWhiteSpaceChar;
 
-  ImportCharSet::m_Ascii['('] |= ImportCharSet::c822SpecialChar;
-  ImportCharSet::m_Ascii[')'] |= ImportCharSet::c822SpecialChar;
-  ImportCharSet::m_Ascii['<'] |= ImportCharSet::c822SpecialChar;
-  ImportCharSet::m_Ascii['>'] |= ImportCharSet::c822SpecialChar;
-  ImportCharSet::m_Ascii['@'] |= ImportCharSet::c822SpecialChar;
-  ImportCharSet::m_Ascii[','] |= ImportCharSet::c822SpecialChar;
-  ImportCharSet::m_Ascii[';'] |= ImportCharSet::c822SpecialChar;
-  ImportCharSet::m_Ascii[':'] |= ImportCharSet::c822SpecialChar;
-  ImportCharSet::m_Ascii['\\'] |= ImportCharSet::c822SpecialChar;
-  ImportCharSet::m_Ascii['"'] |= ImportCharSet::c822SpecialChar;
-  ImportCharSet::m_Ascii['.'] |= ImportCharSet::c822SpecialChar;
-  ImportCharSet::m_Ascii['['] |= ImportCharSet::c822SpecialChar;
-  ImportCharSet::m_Ascii[']'] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t('(')] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t(')')] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t('<')] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t('>')] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t('@')] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t(',')] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t(';')] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t(':')] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t('\\')] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t('"')] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t('.')] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t('[')] |= ImportCharSet::c822SpecialChar;
+  ImportCharSet::m_Ascii[uint8_t(']')] |= ImportCharSet::c822SpecialChar;
 
 
 }
